pyproject.toml
[project]
name = "computer-use-agent"
version = "1.0.0"
description = "macOS desktop automation using Gemini Computer Use API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.49.0",
    "pillow>=12.0.0",
    "pyautogui>=0.9.54",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "termcolor>=3.2.0",
]

[project.scripts]
computer-agent = "computer_use_agent.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
